Relocate 'menu_display_timedate()'
